@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^7.4",
         "ext-iconv": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
@@ -29,12 +29,13 @@
         "symfony/finder": "^4.2|^5.0"
     },
     "require-dev": {
+        "ergebnis/phpstan-rules": "^0.14.0",
         "illuminate/console": "^5.8|^6.0|^7.0",
         "illuminate/support": "^5.8|^6.0|^7.0",
-        "ergebnis/phpstan-rules": "^0.14.0",
         "mockery/mockery": "^1.0",
         "phpstan/phpstan-strict-rules": "^0.12",
         "phpunit/phpunit": "^8.0|^9.0",
+        "rector/rector": "^0.7.11",
         "symfony/var-dumper": "^4.2|^5.0",
         "symplify/easy-coding-standard": "^7.1",
         "thecodingmachine/phpstan-strict-rules": "^0.12.0"
