@@ -8,6 +8,7 @@ use Exception;
 use InvalidArgumentException;
 use NunoMaduro\PhpInsights\Application\Console\Contracts\Formatter;
 use NunoMaduro\PhpInsights\Domain\Contracts\HasDetails;
+use NunoMaduro\PhpInsights\Domain\Details;
 use NunoMaduro\PhpInsights\Domain\DetailsComparator;
 use NunoMaduro\PhpInsights\Domain\Insights\Insight;
 use NunoMaduro\PhpInsights\Domain\Insights\InsightCollection;
@@ -19,8 +20,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class Json implements Formatter
 {
-    /** @var OutputInterface */
-    private $output;
+    private OutputInterface $output;
 
     public function __construct(InputInterface $input, OutputInterface $output)
     {
@@ -30,7 +30,6 @@ final class Json implements Formatter
     /**
      * Format the result to the desired format.
      *
-     * @param \NunoMaduro\PhpInsights\Domain\Insights\InsightCollection $insightCollection
      * @param array<int, string> $metrics
      *
      * @throws Exception
@@ -66,7 +65,6 @@ final class Json implements Formatter
     /**
      * Outputs the issues errors according to the format.
      *
-     * @param \NunoMaduro\PhpInsights\Domain\Insights\InsightCollection $insightCollection
      * @param array<string> $metrics
      *
      * @return array<string, array<int, array<string, int|string>>|null>
@@ -105,7 +103,7 @@ final class Json implements Formatter
                 $details = $insight->getDetails();
                 usort($details, $detailsComparator);
 
-                /** @var \NunoMaduro\PhpInsights\Domain\Details $detail */
+                /** @var Details $detail */
                 foreach ($details as $detail) {
                     $current[] = array_filter([
                         'title' => $insight->getTitle(),
