@@ -14,21 +14,12 @@ use Symfony\Component\Console\Style\SymfonyStyle;
  */
 final class Style extends SymfonyStyle
 {
-    /**
-     * @var \Symfony\Component\Console\Input\InputInterface
-     */
-    private $input;
+    private InputInterface $input;
 
-    /**
-     * @var \Symfony\Component\Console\Output\OutputInterface
-     */
-    private $output;
+    private OutputInterface $output;
 
     /**
      * Style constructor.
-     *
-     * @param \Symfony\Component\Console\Input\InputInterface $input
-     * @param \Symfony\Component\Console\Output\OutputInterface $output
      */
     public function __construct(InputInterface $input, OutputInterface $output)
     {
@@ -37,16 +28,12 @@ final class Style extends SymfonyStyle
 
     /**
      * Waits for Enter key.
-     *
-     * @param string $category
-     *
-     * @return \NunoMaduro\PhpInsights\Application\Console\Style
      */
     public function waitForKey(string $category): Style
     {
         $stdin = fopen('php://stdin', 'r');
 
-        if ($stdin !== false && $this->output instanceof ConsoleOutput && $this->input->isInteractive() === true) {
+        if ($stdin !== false && $this->output instanceof ConsoleOutput && $this->input->isInteractive()) {
             $this->newLine();
             $section = $this->output->section();
             $section->writeln(sprintf('<title>Press enter to see %s issues...</title>', strtolower($category)));
