@@ -11,6 +11,7 @@ use NunoMaduro\PhpInsights\Domain\Details;
 use NunoMaduro\PhpInsights\Domain\Helper\Files;
 use PhpCsFixer\Fixer\FixerInterface;
 use PhpCsFixer\Tokenizer\Tokens;
+use SplFileInfo;
 
 /**
  * Decorates original php-cs-fixers with additional behavior.
@@ -21,24 +22,17 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
 {
     use FixPerFileCollector;
 
-    /**
-     * @var \PhpCsFixer\Fixer\FixerInterface
-     */
-    private $fixer;
+    private FixerInterface $fixer;
     /**
      * @var array<string, \Symfony\Component\Finder\SplFileInfo>
      */
-    private $exclude;
+    private array $exclude;
     /**
      * @var array<\NunoMaduro\PhpInsights\Domain\Details>
      */
-    private $errors = [];
+    private array $errors = [];
 
     /**
-     * FixerDecorator constructor.
-     *
-     * @param \PhpCsFixer\Fixer\FixerInterface $fixer
-     * @param string $dir
      * @param array<string> $exclude
      */
     public function __construct(FixerInterface $fixer, string $dir, array $exclude)
@@ -60,7 +54,7 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
         return $this->fixer->isRisky();
     }
 
-    public function fix(\SplFileInfo $file, Tokens $tokens): void
+    public function fix(SplFileInfo $file, Tokens $tokens): void
     {
         if ($this->skipFilesFromExcludedFiles($file)) {
             return;
@@ -79,7 +73,7 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
         return $this->fixer->getPriority();
     }
 
-    public function supports(\SplFileInfo $file): bool
+    public function supports(SplFileInfo $file): bool
     {
         if ($this->skipFilesFromExcludedFiles($file)) {
             return false;
@@ -90,8 +84,6 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
 
     /**
      * Checks if the insight detects an issue.
-     *
-     * @return bool
      */
     public function hasIssue(): bool
     {
@@ -100,8 +92,6 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
 
     /**
      * Gets the title of the insight.
-     *
-     * @return string
      */
     public function getTitle(): string
     {
@@ -114,8 +104,6 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
 
     /**
      * Get the class name of Insight used.
-     *
-     * @return string
      */
     public function getInsightClass(): string
     {
@@ -139,7 +127,7 @@ final class FixerDecorator implements FixerInterface, InsightContract, HasDetail
         $this->errors[] = Details::make()->setFile($file)->setDiff($diff)->setMessage($diff);
     }
 
-    private function skipFilesFromExcludedFiles(\SplFileInfo $file): bool
+    private function skipFilesFromExcludedFiles(SplFileInfo $file): bool
     {
         $path = $file->getRealPath();
 
