@@ -16,22 +16,15 @@ final class InsightCollection
     /**
      * @var array<string, array<\NunoMaduro\PhpInsights\Domain\Contracts\Insight>>
      */
-    private $insightsPerMetric = [];
+    private array $insightsPerMetric = [];
 
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Collector
-     */
-    private $collector;
+    private Collector $collector;
 
-    /**
-     * @var Results
-     */
-    private $results;
+    private ?Results $results = null;
 
     /**
      * Creates a new instance of the Insight Collection.
      *
-     * @param  \NunoMaduro\PhpInsights\Domain\Collector  $collector
      * @param  array<string, array<\NunoMaduro\PhpInsights\Domain\Contracts\Insight>>  $insightsPerMetric
      */
     public function __construct(Collector $collector, array $insightsPerMetric)
@@ -66,8 +59,6 @@ final class InsightCollection
     /**
      * Gets all insights from given metric.
      *
-     * @param  \NunoMaduro\PhpInsights\Domain\Contracts\Metric  $metric
-     *
      * @return array<\NunoMaduro\PhpInsights\Domain\Contracts\Insight>
      */
     public function allFrom(Metric $metric): array
@@ -77,8 +68,6 @@ final class InsightCollection
 
     /**
      * Returns the results of the code taking in consideration the current insights.
-     *
-     * @return \NunoMaduro\PhpInsights\Domain\Results
      */
     public function results(): Results
     {
@@ -95,9 +84,7 @@ final class InsightCollection
                 $perCategory[$category] = [];
             }
 
-            $perCategory[$category] = array_merge(
-                $perCategory[$category], $insights
-            );
+            $perCategory[$category] = [...$perCategory[$category], ...$insights];
         }
         $this->results = new Results($this->collector, $perCategory);
 
