@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace NunoMaduro\PhpInsights\Domain\Insights;
 
+use InvalidArgumentException;
 use NunoMaduro\PhpInsights\Domain\Analyser;
 use NunoMaduro\PhpInsights\Domain\Configuration;
 use NunoMaduro\PhpInsights\Domain\Contracts\HasInsights;
 use NunoMaduro\PhpInsights\Domain\Contracts\Insight;
 use NunoMaduro\PhpInsights\Domain\Contracts\Repositories\FilesRepository;
 use NunoMaduro\PhpInsights\Domain\Exceptions\DirectoryNotFound;
+use SplFileInfo;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -17,27 +19,14 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class InsightCollectionFactory
 {
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Contracts\Repositories\FilesRepository
-     */
-    private $filesRepository;
+    private FilesRepository $filesRepository;
 
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Analyser
-     */
-    private $analyser;
+    private Analyser $analyser;
 
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Configuration
-     */
-    private $config;
+    private Configuration $config;
 
     /**
      * Creates a new instance of InsightCollection Factory.
-     *
-     * @param \NunoMaduro\PhpInsights\Domain\Contracts\Repositories\FilesRepository $filesRepository
-     * @param \NunoMaduro\PhpInsights\Domain\Analyser $analyser
-     * @param \NunoMaduro\PhpInsights\Domain\Configuration $config
      */
     public function __construct(
         FilesRepository $filesRepository,
@@ -51,9 +40,6 @@ final class InsightCollectionFactory
 
     /**
      * @param array<string> $metrics
-     * @param OutputInterface $consoleOutput
-     *
-     * @return \NunoMaduro\PhpInsights\Domain\Insights\InsightCollection
      */
     public function get(
         array $metrics,
@@ -63,10 +49,11 @@ final class InsightCollectionFactory
         $commonPath = $this->config->getCommonPath();
 
         try {
-            $files = array_map(static function (\SplFileInfo $file) {
+            $files = array_map(static function (SplFileInfo $file) {
                 return $file->getRealPath();
-            }, $this->filesRepository->within($paths, $this->config->getExcludes())->getFiles());
-        } catch (\InvalidArgumentException $exception) {
+            }, $this->filesRepository->within($paths, $this->config->getExcludes())->getFiles()
+            );
+        } catch (InvalidArgumentException $exception) {
             throw new DirectoryNotFound($exception->getMessage(), 0, $exception);
         }
 
@@ -74,22 +61,25 @@ final class InsightCollectionFactory
 
         $insightsClasses = [];
         foreach ($metrics as $metricClass) {
-            $insightsClasses = array_merge($insightsClasses, $this->getInsights($metricClass));
+            $insightsClasses = [...$insightsClasses, ...$this->getInsights($metricClass)];
         }
 
         $insightFactory = new InsightFactory($this->filesRepository, $insightsClasses, $this->config);
         $insightsForCollection = [];
         foreach ($metrics as $metricClass) {
-            $insightsForCollection[$metricClass] = array_map(function (string $insightClass) use ($insightFactory, $collector, $consoleOutput) {
-                if (! array_key_exists(Insight::class, class_implements($insightClass))) {
-                    return $insightFactory->makeFrom(
-                        $insightClass,
-                        $consoleOutput
-                    );
-                }
+            $insightsForCollection[$metricClass] = array_map(
+                function (string $insightClass) use ($insightFactory, $collector, $consoleOutput) {
+                    if (! array_key_exists(Insight::class, class_implements($insightClass))) {
+                        return $insightFactory->makeFrom(
+                            $insightClass,
+                            $consoleOutput
+                        );
+                    }
 
-                return new $insightClass($collector, $this->config->getConfigForInsight($insightClass));
-            }, $this->getInsights($metricClass));
+                    return new $insightClass($collector, $this->config->getConfigForInsight($insightClass));
+                },
+                $this->getInsights($metricClass),
+            );
         }
 
         return new InsightCollection($collector, $insightsForCollection);
@@ -97,8 +87,6 @@ final class InsightCollectionFactory
 
     /**
      * Returns the `Insights` from the given metric class.
-     *
-     * @param string $metricClass
      *
      * @return array<string>
      */
@@ -113,7 +101,7 @@ final class InsightCollectionFactory
         ) ? $metric->getInsights() : [];
 
         $toAdd = $this->config->getAddedInsightsByMetric($metricClass);
-        $insights = array_merge($insights, $toAdd);
+        $insights = [...$insights, ...$toAdd];
 
         // Remove insights based on config.
         return array_diff($insights, $this->config->getRemoves());
