@@ -15,23 +15,19 @@ use NunoMaduro\PhpInsights\Domain\Insights\ForbiddenSecurityIssues;
  */
 final class Results
 {
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Collector
-     */
-    private $collector;
+    private Collector $collector;
 
     /**
      * @var array<string, array<\NunoMaduro\PhpInsights\Domain\Contracts\Insight>>
      */
-    private $perCategoryInsights;
+    private array $perCategoryInsights;
 
     /**
      * Creates a new instance of results.
      *
-     * @param  \NunoMaduro\PhpInsights\Domain\Collector  $collector
      * @param  array<string, array<\NunoMaduro\PhpInsights\Domain\Contracts\Insight>>  $perCategoryInsights
      */
-    public function __construct(\NunoMaduro\PhpInsights\Domain\Collector $collector, array $perCategoryInsights)
+    public function __construct(Collector $collector, array $perCategoryInsights)
     {
         $this->collector = $collector;
         $this->perCategoryInsights = $perCategoryInsights;
@@ -39,8 +35,6 @@ final class Results
 
     /**
      * Gets the code quality.
-     *
-     * @return float
      */
     public function getCodeQuality(): float
     {
@@ -49,8 +43,6 @@ final class Results
 
     /**
      * Gets the code quality.
-     *
-     * @return float
      */
     public function getComplexity(): float
     {
@@ -66,8 +58,6 @@ final class Results
 
     /**
      * Gets the code quality.
-     *
-     * @return float
      */
     public function getStructure(): float
     {
@@ -76,8 +66,6 @@ final class Results
 
     /**
      * Gets the code quality.
-     *
-     * @return float
      */
     public function getDependencies(): float
     {
@@ -86,8 +74,6 @@ final class Results
 
     /**
      * Gets the style quality.
-     *
-     * @return float
      */
     public function getStyle(): float
     {
@@ -96,8 +82,6 @@ final class Results
 
     /**
      * Gets number of security issues.
-     *
-     * @return int
      */
     public function getTotalSecurityIssues(): int
     {
@@ -154,10 +138,6 @@ final class Results
     }
     /**
      * Returns the percentage of the given category.
-     *
-     * @param  string  $category
-     *
-     * @return float
      */
     private function getPercentage(string $category): float
     {
