@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace NunoMaduro\PhpInsights\Infrastructure\Repositories;
 
 use NunoMaduro\PhpInsights\Domain\Contracts\Repositories\FilesRepository;
+use SplFileInfo;
 use Symfony\Component\Finder\Finder;
 
 /**
@@ -12,25 +13,22 @@ use Symfony\Component\Finder\Finder;
  */
 final class LocalFilesRepository implements FilesRepository
 {
-    /**
-     * @var \Symfony\Component\Finder\Finder
-     */
-    private $finder;
+    private Finder $finder;
 
     /**
      * @var array<\Symfony\Component\Finder\SplFileInfo>
      */
-    private $files;
+    private ?array $files = null;
 
     /**
      * @var array<mixed>
      */
-    private $fileList;
+    private ?array $fileList = null;
 
     /**
      * @var array<string>
      */
-    private $directoryList;
+    private ?array $directoryList = null;
 
     public function __construct(Finder $finder)
     {
@@ -86,7 +84,7 @@ final class LocalFilesRepository implements FilesRepository
             ->exclude(['vendor', 'tests', 'Tests', 'test', 'Test'])
             ->notName(['*.blade.php'])
             ->ignoreUnreadableDirs()
-            ->in($this->directoryList)
+            ->in($this->directoryList ?? [])
             ->notPath($exclude);
 
         foreach ($exclude as $value) {
@@ -104,10 +102,14 @@ final class LocalFilesRepository implements FilesRepository
     private function getSingleFiles(): array
     {
         $this->finder = Finder::create()
-            ->in($this->fileList['dirname'])
-            ->name($this->fileList['basename'])
-            ->filter(function (\SplFileInfo $file): bool {
-                return \in_array($file->getPathname(), $this->fileList['full_path'], true);
+            ->in($this->fileList['dirname'] ?? [])
+            ->name($this->fileList['basename'] ?? '')
+            ->filter(function (SplFileInfo $file): bool {
+                return in_array(
+                    $file->getPathname(),
+                    $this->fileList['full_path'] ?? '',
+                    true
+                );
             });
 
         return $this->getFilesList();
