@@ -4,19 +4,13 @@ declare(strict_types=1);
 
 namespace Tests\Fixtures\Domain\Fixer;
 
-use NunoMaduro\PhpInsights\Domain\Analyser;
 use NunoMaduro\PhpInsights\Application\ConfigResolver;
+use NunoMaduro\PhpInsights\Domain\Analyser;
 
 abstract class UnorderedUse
 {
-    /**
-     * @var \NunoMaduro\PhpInsights\Domain\Analyser
-     */
-    protected $analyser;
-    /**
-     * @var \NunoMaduro\PhpInsights\Application\ConfigResolver
-     */
-    protected $configResolver;
+    protected Analyser $analyser;
+    protected ConfigResolver $configResolver;
 
     public function __construct(
         Analyser $analyser,
