@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace NunoMaduro\PhpInsights\Application\Console;
 
+use NunoMaduro\PhpInsights\Application\Console\Contracts\Style;
 use NunoMaduro\PhpInsights\Application\Console\Styles\Bold;
 use NunoMaduro\PhpInsights\Application\Console\Styles\Title;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -16,24 +17,20 @@ final class OutputDecorator
     /**
      * @var array<string>
      */
-    private static $styles = [
+    private static array $styles = [
         Title::class,
         Bold::class,
     ];
 
     /**
      * Decorates the given output with styles.
-     *
-     * @param  \Symfony\Component\Console\Output\OutputInterface  $output
-     *
-     * @return \Symfony\Component\Console\Output\OutputInterface
      */
     public static function decorate(OutputInterface $output): OutputInterface
     {
         foreach (self::$styles as $styleClass) {
             $style = new $styleClass();
 
-            /** @var \NunoMaduro\PhpInsights\Application\Console\Contracts\Style $style */
+            /** @var Style $style */
             $style->addTo($output);
         }
 
