@@ -20,12 +20,8 @@ final class InsightLoaders
     public function __invoke(): array
     {
         return [
-            SniffLoader::class => static function (): SniffLoader {
-                return new SniffLoader();
-            },
-            FixerLoader::class => static function (): FixerLoader {
-                return new FixerLoader();
-            },
+            SniffLoader::class => static fn (): SniffLoader => new SniffLoader(),
+            FixerLoader::class => static fn (): FixerLoader => new FixerLoader(),
         ];
     }
 }
