@@ -39,11 +39,15 @@ final class DetailsComparator
 
     private function functionComparison(Details $first, Details $second): int
     {
-        return ($first->hasFunction() ? $first->getFunction() : null) <=> ($second->hasFunction() ? $second->getFunction() : null);
+        return ($first->hasFunction() ? $first->getFunction() : null)
+            <=>
+            ($second->hasFunction() ? $second->getFunction() : null);
     }
 
     private function messageComparison(Details $first, Details $second): int
     {
-        return ($first->hasMessage() ? $first->getMessage() : null) <=> ($second->hasMessage() ? $second->getMessage() : null);
+        return ($first->hasMessage() ? $first->getMessage() : null)
+            <=>
+            ($second->hasMessage() ? $second->getMessage() : null);
     }
 }
