@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace NunoMaduro\PhpInsights\Domain\Exceptions;
 
+use RuntimeException;
+
 /**
  * @internal
  */
-final class DirectoryNotFound extends \RuntimeException
+final class DirectoryNotFound extends RuntimeException
 {
 }
