@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace NunoMaduro\PhpInsights\Domain\Exceptions;
 
-final class InsightClassNotFound extends \RuntimeException
+use RuntimeException;
+
+final class InsightClassNotFound extends RuntimeException
 {
 }
