@@ -11,7 +11,10 @@ final class Globally implements HasPercentage
 {
     public function getPercentage(Collector $collector): float
     {
-        $value = count($collector->getFiles()) - $collector->getClasses() - $collector->getInterfaces() - count($collector->getTraits());
+        $value = count($collector->getFiles())
+            - $collector->getClasses()
+            - $collector->getInterfaces()
+            - count($collector->getTraits());
 
         return count($collector->getFiles()) > 0 ? $value / count($collector->getFiles()) * 100 : 0;
     }
